+++ tests/run.py
@@ -88,6 +88,7 @@ _TEST_MODULES = [
     "tests.test_vl53l0x",
     "tests.test_vl53l1x",
     "tests.test_example_pins",
+    "tests.test_parameters",
     "tests.test_line_follow",
     "tests.test_qtr",
     "tests.test_qtr_align",